refactor: puzzle output through singular solve method [incremental]

+++ src/main/java/com/gotreaux/aoc/puzzles/year2020/day2/PositionPasswordPolicy.java
@@ -1,7 +1,6 @@
 package com.gotreaux.aoc.puzzles.year2020.day2;
 
-public record PositionPasswordPolicy(Long first, Long second, char target)
-        implements PasswordPolicy {
+public record PositionPasswordPolicy(int first, int second, char target) implements PasswordPolicy {
     public PositionPasswordPolicy {
         if (first < 0 || second <= 0) {
             throw new IllegalArgumentException("Password policy should define positive position!");
@@ -10,7 +9,6 @@ public record PositionPasswordPolicy(Long first, Long second, char target)
 
     @Override
     public boolean passes(String password) {
-        return password.charAt(first.intValue() - 1) == target
-                ^ password.charAt(second.intValue() - 1) == target;
+        return password.charAt(first - 1) == target ^ password.charAt(second - 1) == target;
     }
 }
